hm1
collected the first homework in directory
