+++ app/src/main/java/com/example/androidlab01/homework_1/ThirdActivity.kt
@@ -1,4 +1,4 @@
-package com.example.androidlab01
+package com.example.androidlab01.homework_1
 
 import android.content.Context
 import android.os.Bundle
@@ -18,25 +18,25 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-class SecondActivity : ComponentActivity() {
+import com.example.androidlab01.R
+
+class ThirdActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
         setContent {
-            val DEFAULT_TEXT = stringResource(R.string.default_second_screen_text)
-            val isTextFromMainExist = intent.getStringExtra("textFromMain") != null
-            val secondScreenText = intent.getStringExtra("textFromMain") ?: DEFAULT_TEXT
+            val textFromMain = intent.getStringExtra("textFromMain") ?: stringResource(R.string.default_third_screen_text)
             val context = LocalContext.current
 
-            SecondScreen(secondScreenText, context, isTextFromMainExist)
+            ThirdScreen(textFromMain, context)
         }
     }
 }
 
 @Composable
-fun SecondScreen(
-    secondScreenText: String,
-    context: Context,
-    isTextfromMainExist: Boolean
+fun ThirdScreen(
+    textFromMain: String,
+    context: Context
 ) {
     Column (
         modifier = Modifier.fillMaxHeight(0.8f).fillMaxWidth(),
@@ -44,28 +44,15 @@ fun SecondScreen(
         horizontalAlignment = Alignment.CenterHorizontally
     ){
         Text(
-            text = secondScreenText,
+            text = textFromMain,
             fontSize = 25.sp,
             modifier = Modifier.padding(16.dp)
         )
         Button(onClick = {
             navigateToMainActivity(context)
-        },
-            modifier = Modifier.padding(16.dp)
-        ) {
-            Text(text = stringResource(R.string.button_to_main), fontSize = 25.sp)
-        }
-        Button(
-            onClick = {
-                navigateToOthersActivity(
-                    context,
-                    ThirdActivity::class.java,
-                    if (isTextfromMainExist) secondScreenText else ""
-                )
-            },
-            modifier = Modifier.padding(16.dp)
-        ) {
-            Text(text = stringResource(R.string.button_to_third), fontSize = 25.sp)
+        }) {
+            Text(text = stringResource(R.string.button_to_main),
+                fontSize = 25.sp)
         }
     }
 }
